Module google_consumersurveys2::api

Structs

ConsumerSurveys
Central instance to access all ConsumerSurveys related resource activities

FieldMask
There is no detailed description.

MobileAppPanel
There is no detailed description.

MobileAppPanelsListResponse
There is no detailed description.

MobileapppanelGetCall
Retrieves a MobileAppPanel that is available to the authenticated user.

MobileapppanelListCall
Lists the MobileAppPanels available to the authenticated user.

MobileapppanelMethods
A builder providing access to all methods supported on mobileapppanel resources. It is not used directly, but through the ConsumerSurveys hub.

MobileapppanelUpdateCall
Updates a MobileAppPanel. Currently the only property that can be updated is the owners property.

PageInfo
There is no detailed description.

ResultGetCall
Retrieves any survey results that have been produced so far. Results are formatted as an Excel file. You must add "?alt=media" to the URL as an argument to get results.

ResultMethods
A builder providing access to all methods supported on result resources. It is not used directly, but through the ConsumerSurveys hub.

ResultsGetRequest
There is no detailed description.

ResultsMask
There is no detailed description.

Survey
There is no detailed description.

SurveyAudience
There is no detailed description.

SurveyCost
There is no detailed description.

SurveyDeleteCall
Removes a survey from view in all user GET requests.

SurveyGetCall
Retrieves information about the specified survey.

SurveyInsertCall
Creates a survey.

SurveyListCall
Lists the surveys owned by the authenticated user.

SurveyMethods
A builder providing access to all methods supported on survey resources. It is not used directly, but through the ConsumerSurveys hub.

SurveyQuestion
There is no detailed description.

SurveyQuestionImage
There is no detailed description.

SurveyRejection
There is no detailed description.

SurveyResults
There is no detailed description.

SurveyStartCall
Begins running a survey.

SurveyStopCall
Stops a running survey.

SurveyUpdateCall
Updates a survey. Currently the only property that can be updated is the owners property.

SurveysDeleteResponse
There is no detailed description.

SurveysListResponse
There is no detailed description.

SurveysStartRequest
There is no detailed description.

SurveysStartResponse
There is no detailed description.

SurveysStopResponse
There is no detailed description.

TokenPagination
There is no detailed description.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.