Struct google_consumersurveys2::Survey

pub struct Survey {
    pub owners: Option<Vec<String>>,
    pub description: Option<String>,
    pub title: Option<String>,
    pub customer_data: Option<String>,
    pub state: Option<String>,
    pub audience: Option<SurveyAudience>,
    pub cost: Option<SurveyCost>,
    pub survey_url_id: Option<String>,
    pub questions: Option<Vec<SurveyQuestion>>,
    pub wanted_response_count: Option<i32>,
}

Representation of an individual survey object.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert surveys (request|response)

list surveys (none)

update surveys (request|response)

get surveys (response)

stop surveys (none)

start surveys (none)

Fields

owners

List of email addresses for survey owners. Must contain at least the address of the user making the API call.

description

Text description of the survey.

title

Optional name that will be given to the survey.

customer_data

Additional information to store on behalf of the API consumer and associate with this question. This binary blob is treated as opaque. This field is limited to 64K bytes.

state

State that the survey is in. Can be modified to start, stop, or pause survey.

audience

Targeting-criteria message containing demographic information

cost

Cost to run the survey and collect the necessary number of responses.

survey_url_id

Unique survey ID, that is viewable in the URL of the Survey Creator UI

questions

List of questions defining the survey.

wanted_response_count

Number of responses desired for the survey.