Struct google_consumersurveys2::SurveyMethods

pub struct SurveyMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on survey resources. It is not used directly, but through the Consumersurveys hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_consumersurveys2 as consumersurveys2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use consumersurveys2::Consumersurveys;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Consumersurveys::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)`, `insert(...)`, `list(...)`, `start(...)`, `stop(...)` and `update(...)`
// to build up your call.
let rb = hub.surveys();

Methods

impl<'a, C, A> SurveyMethods<'a, C, A>

fn insert(&self, request: Survey) -> SurveyInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a survey.

Arguments

request - No description provided.

fn list(&self) -> SurveyListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the surveys owned by the authenticated user.

fn update(&self, request: Survey, survey_url_id: &str) -> SurveyUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a survey. Currently the only property that can be updated is the owners property.

Arguments

request - No description provided.

surveyUrlId - External URL ID for the survey.

fn get(&self, survey_url_id: &str) -> SurveyGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves information about the specified survey.

Arguments

surveyUrlId - External URL ID for the survey.

fn stop(&self, resource_id: &str) -> SurveyStopCall<'a, C, A>

Create a builder to help you perform the following task:

Stops a running survey.

Arguments

resourceId - No description provided.

fn start(&self, resource_id: &str) -> SurveyStartCall<'a, C, A>

Create a builder to help you perform the following task:

Begins running a survey.

Arguments

resourceId - No description provided.